==> R/7_Final_models.R <==
# Credits ---------------------------

# Script created by
# Bruno M. Carvalho (https://github.com/brunomc-eco)
#  & Luara Tourinho (https://github.com/luaratourinho)

# Edited by
# Julia Niemeyer & Luara Tourinho
# Last update: 14 feb 2022


##Running ENM for multiple species & GCMs
# Required packages
library(parallel)
library(foreach)
library(doParallel)
require(raster)
library(rgeos)
library(rgdal)
library(dismo)
library(rJava)
library(kernlab)
library(randomForest)
library(maptools)
library(SDMTools)
library(dplyr)
library(biomod2)
library(beepr)
library(maxnet)



indexOf <- function(v,findFor) {
  i=0
  for(i2 in v) {
    i = i + 1
    if (i2==findFor){
      return(i)
    }
  }
  return(0)
}


#setwd("/Mydirectory")



# Settings ----------------------------------------------------------------
#########           ACTION NEEDED        ##########

# for GLM, RandomForest and SVM
##Set your model to the biovariables you selected. It should be exactly like the names of the environmental rasters.
model <- pa ~bio_2+bio_3+bio_10+bio_15+bio_18

# partitions (4 = 75% train e 25% test; 5 = 80% train e 20% test; 10 = 90% train e 10% test)
k = 10
# number of background
bg.pt = 10000
# threshold ('spec_sens' = max. spec + sens)
t.met = 'spec_sens'
# Minimum TSS value for ensemble
tss.lim = 0.6

cont.maps = T # save continuous maps by algorithm
bin.maps = T # save binary maps by algorithm
ens.maps = T # save ensemble maps by algorithm

#### Name of .csv file with species occurrences, columns: 'species', 'lon', 'lat'
## Entrar com a planilha limpa pós spthin
file = "./outputs/03_clean_df_thin_5.csv"

##minimum occurrence records to run analysis
n_min <- 15


BCC <- "bcc" 
IPSL <- "ipsl"
MIROC <-"miroc"


########## END OF ACTION  NEEDED ############

# Reading files -----------------------------------------------------------
sp <- read.table(file, header=TRUE, sep=",")
sp_names <- unique(sp$species)
#sp_names <- sp_names[1:12]

##to test for few species use the following line
#length(sp_names) <- 100

for (a in 1:length(sp_names)){
sp.n = sp_names[[a]]
 message("starting the analysis for ", paste0(sp.n))

presences <- read.csv(file)  %>%
  filter(species == paste0(sp.n)) %>%
  select(species, lon, lat)

if (nrow(presences) < n_min){ ##Will not analyze species with less than 15 occurences
  print('species has less than 15 records and will not be analyzed')
  next
}


# running for one species
sp.names <- as.character(unique(presences$species))

#Read predictor variables (i.e. present maps cropped by mcp)
#raster_files <- list.files('./Maps/Present', full.names = T, 'tif$|bil$')
raster_files <- list.files(paste0("./outputs/", sp.n, "/Pres_env_crop"), full.names = T, 'tif$|bil$')
head(raster_files)
predictors <- stack(raster_files)

# Read your future environmental rasters selected by correlation
#raster_files2 <- list.files('./Maps/Future/rcp45', full.names = T, 'tif$|bil$')

## RCP OR GCM 1
raster_files2 <- list.files(paste0("./outputs/", sp.n, "/Fut_env_crop/", BCC), full.names = T, 'tif$|bil$')
head(raster_files2)
future_variable <- stack(raster_files2)
#future_variable[[3]] <- future_variable[[3]]/10 ##divide Temperature values by 10 [https://worldclim.org/data/v1.4/formats.html]
#future_variable[[4]] <- future_variable[[4]]/10
#names(future_variable)


#### RCP OR GCM 2
raster_files3 <- list.files(paste0("./outputs/", sp.n, "/Fut_env_crop/", IPSL), full.names = T, 'tif$|bil$')
head(raster_files3)
future_variable2 <- stack(raster_files3)
#future_variable2[[3]] <- future_variable2[[3]]/10 ##divide Temperature values by 10 [https://worldclim.org/data/v1.4/formats.html]
#future_variable2[[4]] <- future_variable2[[4]]/10
#names(future_variable)

#### RCP OR GCM 3
raster_files4 <- list.files(paste0("./outputs/", sp.n, "/Fut_env_crop/", MIROC), full.names = T, 'tif$|bil$')
head(raster_files4)
future_variable3 <- stack(raster_files4)
#future_variable3[[3]] <- future_variable3[[3]]/10 ##divide Temperature values by 10 [https://worldclim.org/data/v1.4/formats.html]
#future_variable3[[4]] <- future_variable3[[4]]/10
#names(future_variable)

# ENM ---------------------------------------------------------------------


ini = Sys.time()
pb <- txtProgressBar(min = 1, max = length(sp.names)+1, style = 3)

# Number of occurrences to perform pseudoabsence sampling
#Tive que add o plyr::count aqui pra não confundir com dplyr senão dá erro
lim = plyr::count(presences$species)$freq[indexOf(plyr::count(presences$species)$x, sp.n)]

started_time = Sys.time()
cat( format( started_time, "%a %b %d %X %Y"), '-', 'STARTED', '\n')
cat( format( started_time, "%a %b %d %X %Y"), '-', 'Preparing train and test datasets for', sp.n, 'with ', lim, 'lines...', '\n')

##A pasta final onde tudo vai ser salvo (no loop é dentro da pasta de cada espécie_results talvez)
target_dir = paste(paste0("./outputs/", sp.n, "/results/", sep=""))
dir.create( target_dir )

if(file.exists(paste(target_dir, '/STARTED.txt', sep="")))
  stop("You MUST DELETE previous results folder before continue")

write(format( started_time, "%a %b %d %X %Y"), file=paste(target_dir, 'STARTED.txt', sep=""))

# For using different number of pseudoabsence in each algorithm, for example:
## pseudoausencia = n*100
sp.data <- read.csv(paste0("./outputs/", sp.n,"/pres_pseudoabs.csv"), header=TRUE, sep=',')

#colocar aqui embaixo  outra planilha depois de rodar pro outro modelo
##pseudoausencia = pres
# To random forest
sp.data2 <- read.csv(paste0("./outputs/", sp.n,"/pres_pseudoabs2.csv"), header=TRUE, sep=',')


# For bioclim and Maxent:
pres <- sp.data[sp.data$pa==1,2:3]
abs <- sp.data[sp.data$pa!=1,2:3]
bg <- randomPoints(predictors, bg.pt)
colnames(bg) <- c("lon", "lat")


## Crossvalidation partition in 10 groups (são 10 partições)
set.seed(10)
foldpres <- kfold(pres, 4)
set.seed(10)
foldabs <- kfold(abs, 4)

prestrain <- list()
prestest <- list()
abstrain <- list()
abstest <- list()
for(i in 1:k){
  foldpres <- kfold(pres, 4)
  foldabs <- kfold(abs, 4)
  prestrain[[i]] <- pres[foldpres != 1,]
  prestest[[i]] <- pres[foldpres == 1,]
  abstrain[[i]] <- abs[foldabs != 1,]
  abstest[[i]] <- abs[foldabs == 1,]
}


# For SVM, GLM Maxnet:
train <- list()
pa_train <- list()
predtrain <- list()
testpres <- list()
testabs <- list()
data <- list()
p <- list()

for(i in 1:k){
  train[[i]] <- rbind(prestrain[[i]], abstrain[[i]])
  pa_train[[i]] <- c(rep(1, nrow(prestrain[[i]])), rep(0, nrow(abstrain[[i]])))
  predtrain[[i]] <- raster::extract(predictors, train[[i]])
  predtrain[[i]] <- na.omit(data.frame(cbind(pa=pa_train[[i]], predtrain[[i]])))
  data[[i]] <- predtrain[[i]][,-1]
  p[[i]] <- predtrain[[i]][,1]
}


#For Random Forest
pres2 <- sp.data2[sp.data2$pa==1,2:3]
abs2 <- sp.data2[sp.data2$pa!=1,2:3]
set.seed(10)
foldpres2 <- kfold(pres2, 4)
set.seed(10)
foldabs2 <- kfold(abs2, 4)

prestrain2 <- list()
prestest2 <- list()
abstrain2 <- list()
abstest2 <- list()
for(i in 1:k){
  foldpres2 <- kfold(pres2, 4)
  foldabs2 <- kfold(abs2, 4)
  prestrain2[[i]] <- pres2[foldpres2 != 1,]
  prestest2[[i]] <- pres2[foldpres2 == 1,]
  abstrain2[[i]] <- abs2[foldabs2 != 1,]
  abstest2[[i]] <- abs2[foldabs2 == 1,]
}

train2 <- list()
pa_train2 <- list()
predtrain2 <- list()
testpres2 <- list()
testabs2 <- list()

for(i in 1:k){
  train2[[i]] <- rbind(prestrain2[[i]], abstrain2[[i]])
  pa_train2[[i]] <- c(rep(1, nrow(prestrain2[[i]])), rep(0, nrow(abstrain2[[i]])))
  predtrain2[[i]] <- raster::extract(predictors, train2[[i]])
  predtrain2[[i]] <- data.frame(cbind(pa=pa_train2[[i]], predtrain2[[i]]))
}


# For xxx:
# pres3 <- sp.data[sp.data$pa==1,2:3]
# abs3 <- sp.data[sp.data$pa!=1,2:3]
# set.seed(10)
# foldpres3 <- kfold(pres3, 4)
# set.seed(10)
# foldabs3 <- kfold(abs3, 4)
# 
# prestrain3 <- list()
# prestest3 <- list()
# abstrain3 <- list()
# abstest3 <- list()
# for(i in 1:k){
#   foldpres3 <- kfold(pres3, 4)
#   foldabs3 <- kfold(abs3, 4)
#   prestrain3[[i]] <- pres3[foldpres3 != 1,]
#   prestest3[[i]] <- pres3[foldpres3 == 1,]
#   abstrain3[[i]] <- abs3[foldabs3 != 1,]
#   abstest3[[i]] <- abs3[foldabs3 == 1,]
# }
# 
# train3 <- list()
# pa_train3 <- list()
# predtrain3 <- list()
# testpres3 <- list()
# testabs3 <- list()
# 
# for(i in 1:k){
#   train3[[i]] <- rbind(prestrain3[[i]], abstrain3[[i]])
#   pa_train3[[i]] <- c(rep(1, nrow(prestrain3[[i]])), rep(0, nrow(abstrain3[[i]])))
#   predtrain3[[i]] <- raster::extract(predictors, train3[[i]])
#   predtrain3[[i]] <- data.frame(cbind(pa=pa_train3[[i]], predtrain3[[i]]))
# }



# Running models ----------------------------------------------------------

cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running Bioclim model for', sp.n, '...', '\n')
bc <- list()
evbc <- list()
bcTSS <- list()
bcAUC <- list()
bckappa <- list()
bcthres <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running Bioclim (', i, ') model for', sp.n, '...', '\n')
  bc[[i]] <- bioclim(predictors, prestrain[[i]])
  evbc[[i]] <- dismo::evaluate(prestest[[i]], abstest[[i]], bc[[i]], predictors)
  bcTSS[[i]] <- max(evbc[[i]]@TPR + evbc[[i]]@TNR)-1
  bcAUC[[i]] <- evbc[[i]]@auc
  bckappa[[i]] <- max(evbc[[i]]@kappa)
  bcthres[[i]] <- threshold(evbc[[i]], t.met)
}

cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running GLM (logistic regression) for', sp.n, '...', '\n')
gm <- list()
evgm <- list()
gmTSS <- list()
gmAUC <- list()
gmkappa <- list()
gmthres <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running GLM (logistic regression) (', i, ') for', sp.n, '...', '\n')
  gm[[i]] <- glm(model, family=binomial(link="logit"), data=predtrain[[i]])
  evgm[[i]] <- dismo::evaluate(prestest[[i]], abstest[[i]], gm[[i]], predictors)
  gmTSS[[i]] <- max(evgm[[i]]@TPR + evgm[[i]]@TNR)-1
  gmAUC[[i]] <- evgm[[i]]@auc
  gmkappa[[i]] <- max(evgm[[i]]@kappa)
  gmthres[[i]] <- threshold(evgm[[i]], t.met)
}

cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running Random Forest model for', sp.n, '...', '\n')
rf <- list()
evrf <- list()
rfTSS <- list()
rfAUC <- list()
rfkappa <- list()
rfthres <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running Random Forest (', i, ') model for', sp.n, '...', '\n')
  rf[[i]] <- randomForest(model, data=predtrain2[[i]], na.action=na.omit)
  evrf[[i]] <- dismo::evaluate(prestest2[[i]], abstest2[[i]], rf[[i]], predictors)
  rfTSS[[i]] <- max(evrf[[i]]@TPR + evrf[[i]]@TNR)-1
  rfAUC[[i]] <- evrf[[i]]@auc
  rfkappa[[i]] <- max(evrf[[i]]@kappa)
  rfthres[[i]] <- threshold(evrf[[i]], t.met)
}

cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running Maxent model for', sp.n, '...', '\n')
mx <- list()
evmx <- list()
mxTSS <- list()
mxAUC <- list()
mxkappa <- list()
mxthres <- list()


for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running Maxent (', i, ') model for', sp.n, '...', '\n')
  mx[[i]] <- dismo::maxent(predictors, prestrain[[i]], a=bg) #run maxent with java.
  #mx[[i]] <- maxnet(p = p[[i]], data = data[[i]]) #run maxnet without java. You can use whichever you prefer, maxent or maxnet. You must only erase # from one and add on the other.
  evmx[[i]] <- dismo::evaluate(prestest[[i]], abstest[[i]], mx[[i]], predictors)
  mxTSS[[i]] <- max(evmx[[i]]@TPR + evmx[[i]]@TNR)-1
  mxAUC[[i]] <- evmx[[i]]@auc
  mxkappa[[i]] <- max(evmx[[i]]@kappa)
  mxthres[[i]] <- threshold(evmx[[i]], t.met)
}

cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running SVM model for', sp.n, '...', '\n')
sv <- list()
evsv <- list()
svTSS <- list()
svAUC <- list()
svkappa <- list()
svthres <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Running SVM (', i, ') model for', sp.n, '...', '\n')
  sv[[i]] <- ksvm(model, data=predtrain[[i]])
  evsv[[i]] <- dismo::evaluate(prestest[[i]], abstest[[i]], sv[[i]], predictors)
  svTSS[[i]] <- max(evsv[[i]]@TPR + evsv[[i]]@TNR)-1
  svAUC[[i]] <- evsv[[i]]@auc
  svkappa[[i]] <- max(evsv[[i]]@kappa)
  svthres[[i]] <- threshold(evsv[[i]], t.met)
}



# Validation table --------------------------------------------------------


cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Generating Validation table for models of', sp.n, '...', '\n')
bcTSSval <- unlist(bcTSS)
#bcTSSval <- data.frame(matrix(unlist(bcTSS), nrow=length(bcTSS), byrow=TRUE), stringsAsFactors=FALSE)

gmTSSval <- unlist(gmTSS)
#gmTSSval <- data.frame(matrix(unlist(gmTSS), nrow=length(gmTSS), byrow=TRUE), stringsAsFactors=FALSE)

rfTSSval <- unlist(rfTSS)
#rfTSSval <- data.frame(matrix(unlist(rfTSS), nrow=length(rfTSS), byrow=TRUE,), stringsAsFactors=FALSE)

mxTSSval <- unlist(mxTSS)
#mxTSSval <- data.frame(matrix(unlist(mxTSS), nrow=length(mxTSS), byrow=TRUE), stringsAsFactors=FALSE)

svTSSval <- unlist(svTSS)
bcAUCval <- unlist(bcAUC)
gmAUCval <- unlist(gmAUC)
rfAUCval <- unlist(rfAUC)
mxAUCval <- unlist(mxAUC)
svAUCval <- unlist(svAUC)
bckappaval <- unlist(bckappa)
gmkappaval <- unlist(gmkappa)
rfkappaval <- unlist(rfkappa)
mxkappaval <- unlist(mxkappa)
svkappaval <- unlist(svkappa)
mod.names <- c(rep ('bc', k), rep('gm', k), rep('rf', k), rep('mx', k), rep('sv', k))
mod.sp <- c(rep(sp.n, k*5)) #conferir a planilha

TSS <- c(bcTSSval, gmTSSval, rfTSSval, mxTSSval, svTSSval)
AUC <- c(bcAUCval, gmAUCval, rfAUCval, mxAUCval, svAUCval)
kappa <- c(bckappaval, gmkappaval, rfkappaval, mxkappaval, svkappaval)
Valid <- data.frame(mod.sp, mod.names, TSS, AUC, kappa, stringsAsFactors=FALSE)

##Pasta de resultados na pasta de cada espécie
write.csv(Valid, file = paste( target_dir, 'Valid_', sp.n, '.csv', sep=""))


### Test if at least one model is valid. If no models are valid it will stop here and continue on the next species

if(all(TSS < tss.lim)){
  print('no models are valid') ##Stop this

  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'No models were valid for', sp.n, 'with ', lim, 'lines...', '\n')

  save.image("./outputs/my_analysis.rData")

  finished_time = Sys.time()
  cat( format( finished_time, "%a %b %d %X %Y"), '-', 'FINISHED', '\n')
  write(format( finished_time, "%a %b %d %X %Y"), file = paste0(target_dir, "FINISHED.txt", sep=""))

  next
}

# Projecting models -------------------------------------------------------



# Projecting Bioclim -------------------------------------------------------

#Drop bad models, project under current and future conditions
cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting Bioclim model of', sp.n, '...', '\n')
cur.bc <- list()
cur.bc.bin <- list()
future_variable.bc <- list()
future_variable.bc.bin <- list()
future_variable2.bc <- list()
future_variable2.bc.bin <- list()
future_variable3.bc <- list()
future_variable3.bc.bin <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting Bioclim (', i, ') model of', sp.n, '...', '\n')
  if(bcTSS[[i]] >= tss.lim){
    cur.bc[[i]] <- predict(predictors, bc[[i]])
    cur.bc.bin[[i]] <- cur.bc[[i]] > bcthres[[i]]
    future_variable.bc[[i]] <- predict(future_variable, bc[[i]])
    future_variable.bc.bin[[i]] <- future_variable.bc[[i]] > bcthres[[i]]
    future_variable2.bc[[i]] <- predict(future_variable2, bc[[i]])
    future_variable2.bc.bin[[i]] <- future_variable2.bc[[i]] > bcthres[[i]]
    future_variable3.bc[[i]] <- predict(future_variable3, bc[[i]])
    future_variable3.bc.bin[[i]] <- future_variable3.bc[[i]] > bcthres[[i]]

  } else {
    cur.bc[[i]] <- NULL
    cur.bc.bin[[i]] <- NULL
    future_variable.bc[[i]] <- NULL
    future_variable.bc.bin[[i]] <- NULL
    future_variable2.bc[[i]] <- NULL
    future_variable2.bc.bin[[i]] <- NULL
    future_variable3.bc[[i]] <- NULL
    future_variable3.bc.bin[[i]] <- NULL
  }
}

##########
cur.bc <- Filter(Negate(is.null), cur.bc) # remove null rasters
cur.bc.bin <- Filter(Negate(is.null), cur.bc.bin)

#Ensemble of binary models (Majority Rule)
cur.bc.ens <- Reduce('+', cur.bc.bin) #sum of pixels
tval <- unique(cur.bc.ens)
tval <- tval[tval != 0]
tval <- median(tval)

if(!is.null(cur.bc.ens)){
  cur.bc.ens.bin <- cur.bc.ens >= tval} else {
    cur.bc.ens.bin <- NULL
  }


##########
future_variable.bc <- Filter(Negate(is.null), future_variable.bc)
future_variable.bc.bin <- Filter(Negate(is.null), future_variable.bc.bin)
future_variable2.bc <- Filter(Negate(is.null), future_variable2.bc)
future_variable2.bc.bin <- Filter(Negate(is.null), future_variable2.bc.bin)
future_variable3.bc <- Filter(Negate(is.null), future_variable3.bc)
future_variable3.bc.bin <- Filter(Negate(is.null), future_variable3.bc.bin)

future_variable.bc.ens <- Reduce('+', future_variable.bc.bin)
tval <- unique(future_variable.bc.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable.bc.ens.bin <- future_variable.bc.ens >= tval

future_variable2.bc.ens <- Reduce('+', future_variable2.bc.bin)
tval <- unique(future_variable2.bc.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable2.bc.ens.bin <- future_variable2.bc.ens >= tval

future_variable3.bc.ens <- Reduce('+', future_variable3.bc.bin)
tval <- unique(future_variable3.bc.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable3.bc.ens.bin <- future_variable3.bc.ens >= tval

##########

# Taking out the zero values and normalizing rasters

if(length(cur.bc) != 0) {
  for(z in 1:length(cur.bc)) {
    adeq = cur.bc[[z]]
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    cur.bc[[z]] <- calc(adeq, adeq_norm)
    }

  for(z in 1:length(future_variable.bc)){
    adeq = future_variable.bc[[z]]
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable.bc[[z]] <- calc(adeq, adeq_norm)
  }

  for(z in 1:length(future_variable2.bc)){
    adeq = future_variable2.bc[[z]]
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable2.bc[[z]] <- calc(adeq, adeq_norm)
  }
  
  for(z in 1:length(future_variable3.bc)){
    adeq = future_variable3.bc[[z]]
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable3.bc[[z]] <- calc(adeq, adeq_norm)
  }

  #Ensemble of continuos models
  x <- stack(cur.bc)
  cur.bc.cont <- calc(x, fun = mean)

  x <- stack(future_variable.bc)
  future_variable.bc.cont <- calc(x, fun = mean)

  x <- stack(future_variable2.bc)
  future_variable2.bc.cont <- calc(x, fun = mean)
  
  x <- stack(future_variable3.bc)
  future_variable3.bc.cont <- calc(x, fun = mean)
  
} else {
  cur.bc.cont <- NULL
  future_variable.bc.cont <- NULL
  future_variable2.bc.cont <- NULL
  future_variable3.bc.cont <- NULL
}

##obs. isso aqui talve remover - ver oq faz mas talvez não precise (rodar e ver)
#for(i in 1:k){
 # min_max <- range(gcmmg.bc[[i]][], na.rm=T)
  #write.table(min_max, paste("gcmmg.bc", i, sp.n, ".txt",  sep="_"))
#}

#rm(cur.bc.bin, cur.bc.cont, cur.bc.ens)
#rm(future_variable.bc.bin, future_variable.bc.cont, future_variable.bc.ens)
###

# Projecting GLM ----------------------------------------------------------


cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting GLMs of', sp.n, '...', '\n')
cur.gm <- list()
cur.gm.bin <- list()
future_variable.gm <- list()
future_variable.gm.bin <- list()
future_variable2.gm <- list()
future_variable2.gm.bin <- list()
future_variable3.gm <- list()
future_variable3.gm.bin <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting GLMs (', i, ') of', sp.n, '...', '\n')
  if(gmTSS[[i]] >= tss.lim){
    cur.gm[[i]] <- predict(predictors, gm[[i]])
    cur.gm.bin[[i]] <- cur.gm[[i]] > gmthres[[i]]
    future_variable.gm[[i]] <- predict(future_variable, gm[[i]])
    future_variable.gm.bin[[i]] <- future_variable.gm[[i]] > gmthres[[i]]
    future_variable2.gm[[i]] <- predict(future_variable2, gm[[i]])
    future_variable2.gm.bin[[i]] <- future_variable2.gm[[i]] > gmthres[[i]]
    future_variable3.gm[[i]] <- predict(future_variable3, gm[[i]])
    future_variable3.gm.bin[[i]] <- future_variable3.gm[[i]] > gmthres[[i]]

  } else {
    cur.gm[[i]] <- NULL
    cur.gm.bin[[i]] <- NULL
    future_variable.gm[[i]] <- NULL
    future_variable.gm.bin[[i]] <- NULL
    future_variable2.gm[[i]] <- NULL
    future_variable2.gm.bin[[i]] <- NULL
    future_variable3.gm[[i]] <- NULL
    future_variable3.gm.bin[[i]] <- NULL
  }
}

##########


cur.gm <- Filter(Negate(is.null), cur.gm)
cur.gm.bin <- Filter(Negate(is.null), cur.gm.bin)

cur.gm.ens <- Reduce('+', cur.gm.bin)
tval <- unique(cur.gm.ens)
tval <- tval[tval != 0]
tval <- median(tval)

if(!is.null(cur.gm.ens)){
cur.gm.ens.bin <- cur.gm.ens >= tval} else {
  cur.gm.ens.bin <- NULL
}

##########
future_variable.gm <- Filter(Negate(is.null), future_variable.gm)
future_variable.gm.bin <- Filter(Negate(is.null), future_variable.gm.bin)
future_variable2.gm <- Filter(Negate(is.null), future_variable2.gm)
future_variable2.gm.bin <- Filter(Negate(is.null), future_variable2.gm.bin)
future_variable3.gm <- Filter(Negate(is.null), future_variable3.gm)
future_variable3.gm.bin <- Filter(Negate(is.null), future_variable3.gm.bin)


future_variable.gm.ens <- Reduce('+', future_variable.gm.bin)
tval <- unique(future_variable.gm.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable.gm.ens.bin <- future_variable.gm.ens >= tval

future_variable2.gm.ens <- Reduce('+', future_variable2.gm.bin)
tval <- unique(future_variable2.gm.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable2.gm.ens.bin <- future_variable2.gm.ens >= tval

future_variable3.gm.ens <- Reduce('+', future_variable3.gm.bin)
tval <- unique(future_variable3.gm.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable3.gm.ens.bin <- future_variable3.gm.ens >= tval


##########

if(length(cur.gm) != 0) {
for(z in 1:length(cur.gm)){
  adeq = cur.gm[[z]]
  minimo <- min(adeq[], na.rm=T)
  maximo <- max(adeq[], na.rm=T)
  adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
  cur.gm[[z]] <- calc(adeq, adeq_norm)
}


for(z in 1:length(future_variable.gm)){
  adeq = future_variable.gm[[z]]
  minimo <- min(adeq[], na.rm=T)
  maximo <- max(adeq[], na.rm=T)
  adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
  future_variable.gm[[z]] <- calc(adeq, adeq_norm)
}

  for(z in 1:length(future_variable2.gm)){
    adeq = future_variable2.gm[[z]]
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable2.gm[[z]] <- calc(adeq, adeq_norm)
  }
  
  for(z in 1:length(future_variable3.gm)){
    adeq = future_variable3.gm[[z]]
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable3.gm[[z]] <- calc(adeq, adeq_norm)
  }

x <- stack(cur.gm)
cur.gm.cont <- calc(x, fun = mean)

x <- stack(future_variable.gm)
future_variable.gm.cont <- calc(x, fun = mean)

x <- stack(future_variable2.gm)
future_variable2.gm.cont <- calc(x, fun = mean)

x <- stack(future_variable3.gm)
future_variable3.gm.cont <- calc(x, fun = mean)

} else {
  cur.gm.cont <- NULL
  future_variable.gm.cont <- NULL
  future_variable2.gm.cont <- NULL
  future_variable3.gm.cont <- NULL
}


# Projecting RF ----------------------------------------------------------



cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting Random Forest models of', sp.n, '...', '\n')
cur.rf <- list()
cur.rf.bin <- list()
future_variable.rf <- list()
future_variable.rf.bin <- list()
future_variable2.rf <- list()
future_variable2.rf.bin <- list()
future_variable3.rf <- list()
future_variable3.rf.bin <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting Random Forest (', i, ') models of', sp.n, '...', '\n')
  if(rfTSS[[i]] >= tss.lim){
    cur.rf[[i]] <- predict(predictors, rf[[i]])
    cur.rf.bin[[i]] <- cur.rf[[i]] > rfthres[[i]]

    future_variable.rf[[i]] <- predict(future_variable, rf[[i]])
    future_variable.rf.bin[[i]] <- future_variable.rf[[i]] > rfthres[[i]]

    future_variable2.rf[[i]] <- predict(future_variable2, rf[[i]])
    future_variable2.rf.bin[[i]] <- future_variable2.rf[[i]] > rfthres[[i]]
    
    future_variable3.rf[[i]] <- predict(future_variable3, rf[[i]])
    future_variable3.rf.bin[[i]] <- future_variable3.rf[[i]] > rfthres[[i]]

  } else {

    cur.rf[[i]] <- NULL
    cur.rf.bin[[i]] <- NULL
    future_variable.rf[[i]] <- NULL
    future_variable.rf.bin[[i]] <- NULL
    future_variable2.rf[[i]] <- NULL
    future_variable2.rf.bin[[i]] <- NULL
    future_variable3.rf[[i]] <- NULL
    future_variable3.rf.bin[[i]] <- NULL

  }
}

##########
cur.rf <- Filter(Negate(is.null), cur.rf)
cur.rf.bin <- Filter(Negate(is.null), cur.rf.bin)

cur.rf.ens <- Reduce('+', cur.rf.bin)
tval <- unique(cur.rf.ens)
tval <- tval[tval != 0]
tval <- median(tval)

if(!is.null(cur.rf.ens)){
  cur.rf.ens.bin <- cur.rf.ens >= tval} else {
    cur.rf.ens.bin <- NULL
  }

##########
future_variable.rf <- Filter(Negate(is.null), future_variable.rf)
future_variable.rf.bin <- Filter(Negate(is.null), future_variable.rf.bin)

future_variable2.rf <- Filter(Negate(is.null), future_variable2.rf)
future_variable2.rf.bin <- Filter(Negate(is.null), future_variable2.rf.bin)

future_variable3.rf <- Filter(Negate(is.null), future_variable3.rf)
future_variable3.rf.bin <- Filter(Negate(is.null), future_variable3.rf.bin)

future_variable.rf.ens <- Reduce('+', future_variable.rf.bin)
tval <- unique(future_variable.rf.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable.rf.ens.bin <- future_variable.rf.ens >= tval

future_variable2.rf.ens <- Reduce('+', future_variable2.rf.bin)
tval <- unique(future_variable2.rf.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable2.rf.ens.bin <- future_variable2.rf.ens >= tval

future_variable3.rf.ens <- Reduce('+', future_variable3.rf.bin)
tval <- unique(future_variable3.rf.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable3.rf.ens.bin <- future_variable3.rf.ens >= tval

##########

if(length(cur.rf) != 0) {
for(z in 1:length(cur.rf)){
  adeq = cur.rf[[z]]
  if(sum(adeq[], na.rm=T)!=0){
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    cur.rf[[z]] <- calc(adeq, adeq_norm)
  }

}


for(z in 1:length(future_variable.rf)){
  adeq = future_variable.rf[[z]]
  if(sum(adeq[], na.rm=T)!=0){
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable.rf[[z]] <- calc(adeq, adeq_norm)
  }
}

for(z in 1:length(future_variable2.rf)){
    adeq = future_variable2.rf[[z]]
    if(sum(adeq[], na.rm=T)!=0){
      minimo <- min(adeq[], na.rm=T)
      maximo <- max(adeq[], na.rm=T)
      adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
      future_variable2.rf[[z]] <- calc(adeq, adeq_norm)
    }
}
  
  for(z in 1:length(future_variable3.rf)){
    adeq = future_variable3.rf[[z]]
    if(sum(adeq[], na.rm=T)!=0){
      minimo <- min(adeq[], na.rm=T)
      maximo <- max(adeq[], na.rm=T)
      adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
      future_variable3.rf[[z]] <- calc(adeq, adeq_norm)
    }
  }

cur.rf <- Filter(Negate(is.null), cur.rf)
future_variable.rf <- Filter(Negate(is.null), future_variable.rf)
future_variable2.rf <- Filter(Negate(is.null), future_variable2.rf)
future_variable3.rf <- Filter(Negate(is.null), future_variable3.rf)

x <- stack(cur.rf)
cur.rf.cont <- calc(x, fun = mean)

x <- stack(future_variable.rf)
future_variable.rf.cont <- calc(x, fun = mean)

x <- stack(future_variable2.rf)
future_variable2.rf.cont <- calc(x, fun = mean)

x <- stack(future_variable3.rf)
future_variable3.rf.cont <- calc(x, fun = mean)

} else {
  cur.rf.cont <- NULL
  future_variable.rf.cont <- NULL
  future_variable2.rf.cont <- NULL
  future_variable3.rf.cont <- NULL
}


# Projecting Maxent --------------------------------------------------------



cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting Maxent models of', sp.n, '...', '\n')
cur.mx <- list()
cur.mx.bin <- list()
future_variable.mx <- list()
future_variable.mx.bin <- list()
future_variable2.mx <- list()
future_variable2.mx.bin <- list()
future_variable3.mx <- list()
future_variable3.mx.bin <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting Maxent (', i, ') models of', sp.n, '...', '\n')
  if(mxTSS[[i]] >= tss.lim){
    cur.mx[[i]] <- predict(predictors, mx[[i]])
    cur.mx.bin[[i]] <- cur.mx[[i]] > mxthres[[i]]

    future_variable.mx[[i]] <- predict(future_variable, mx[[i]])
    future_variable.mx.bin[[i]] <- future_variable.mx[[i]] > mxthres[[i]]

    future_variable2.mx[[i]] <- predict(future_variable2, mx[[i]])
    future_variable2.mx.bin[[i]] <- future_variable2.mx[[i]] > mxthres[[i]]
    
    future_variable3.mx[[i]] <- predict(future_variable3, mx[[i]])
    future_variable3.mx.bin[[i]] <- future_variable3.mx[[i]] > mxthres[[i]]

  } else {
    cur.mx[[i]] <- NULL
    cur.mx.bin[[i]] <- NULL
    future_variable.mx[[i]] <- NULL
    future_variable.mx.bin[[i]] <- NULL
    future_variable2.mx[[i]] <- NULL
    future_variable2.mx.bin[[i]] <- NULL
    future_variable3.mx[[i]] <- NULL
    future_variable3.mx.bin[[i]] <- NULL

  }
}


##########
cur.mx <- Filter(Negate(is.null), cur.mx)
cur.mx.bin <- Filter(Negate(is.null), cur.mx.bin)

cur.mx.ens <- Reduce('+', cur.mx.bin)
tval <- unique(cur.mx.ens)
tval <- tval[tval != 0]
tval <- median(tval)

if(!is.null(cur.mx.ens)){
  cur.mx.ens.bin <- cur.mx.ens >= tval} else {
    cur.mx.ens.bin <- NULL
  }


##########
future_variable.mx <- Filter(Negate(is.null), future_variable.mx)
future_variable.mx.bin <- Filter(Negate(is.null), future_variable.mx.bin)
future_variable2.mx <- Filter(Negate(is.null), future_variable2.mx)
future_variable2.mx.bin <- Filter(Negate(is.null), future_variable2.mx.bin)
future_variable3.mx <- Filter(Negate(is.null), future_variable3.mx)
future_variable3.mx.bin <- Filter(Negate(is.null), future_variable3.mx.bin)

future_variable.mx.ens <- Reduce('+', future_variable.mx.bin)
tval <- unique(future_variable.mx.ens)
tval <- tval[tval != 0]
tval <- median(tval)

if(!is.null(future_variable.mx.ens)){
  future_variable.mx.ens.bin <- future_variable.mx.ens >= tval} else {
    future_variable.mx.ens.bin <- NULL
  }

future_variable2.mx.ens <- Reduce('+', future_variable2.mx.bin)
tval <- unique(future_variable2.mx.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable2.mx.ens.bin <- future_variable2.mx.ens >= tval

if(!is.null(future_variable2.mx.ens)){
  future_variable2.mx.ens.bin <- future_variable2.mx.ens >= tval} else {
    future_variable2.mx.ens.bin <- NULL
  }

future_variable3.mx.ens <- Reduce('+', future_variable3.mx.bin)
tval <- unique(future_variable3.mx.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable3.mx.ens.bin <- future_variable3.mx.ens >= tval

if(!is.null(future_variable3.mx.ens)){
  future_variable3.mx.ens.bin <- future_variable3.mx.ens >= tval} else {
    future_variable3.mx.ens.bin <- NULL
  }

##########

if(length(cur.mx) != 0) {
for(z in 1:length(cur.mx)){
  adeq = cur.mx[[z]]
  if(sum(adeq[], na.rm=T)!=0){
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    cur.mx[[z]] <- calc(adeq, adeq_norm)
  }
}

for(z in 1:length(future_variable.mx)){
  adeq = future_variable.mx[[z]]
  if(sum(adeq[], na.rm=T)!=0){
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable.mx[[z]] <- calc(adeq, adeq_norm)
  }
}

for(z in 1:length(future_variable2.mx)){
    adeq = future_variable2.mx[[z]]
    if(sum(adeq[], na.rm=T)!=0){
      minimo <- min(adeq[], na.rm=T)
      maximo <- max(adeq[], na.rm=T)
      adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
      future_variable2.mx[[z]] <- calc(adeq, adeq_norm)
    }
}
  
  for(z in 1:length(future_variable3.mx)){
    adeq = future_variable3.mx[[z]]
    if(sum(adeq[], na.rm=T)!=0){
      minimo <- min(adeq[], na.rm=T)
      maximo <- max(adeq[], na.rm=T)
      adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
      future_variable3.mx[[z]] <- calc(adeq, adeq_norm)
    }
  }

x <- stack(cur.mx)
cur.mx.cont <- calc(x, fun = mean)

x <- stack(future_variable.mx)
future_variable.mx.cont <- calc(x, fun = mean)

x <- stack(future_variable2.mx)
future_variable2.mx.cont <- calc(x, fun = mean)

x <- stack(future_variable3.mx)
future_variable3.mx.cont <- calc(x, fun = mean)

} else {
  cur.mx.cont <- NULL
  future_variable.mx.cont <- NULL
  future_variable2.mx.cont <- NULL
  future_variable3.mx.cont <- NULL
}


# Projecting SVM ----------------------------------------------------------



cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting SVM models of', sp.n, '...', '\n')
cur.sv <- list()
cur.sv.bin <- list()
future_variable.sv <- list()
future_variable.sv.bin <- list()
future_variable2.sv <- list()
future_variable2.sv.bin <- list()
future_variable3.sv <- list()
future_variable3.sv.bin <- list()

for(i in 1:k){
  cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting SVM (', i, ') models of', sp.n, '...', '\n')
  if(svTSS[[i]] >= tss.lim){
    cur.sv[[i]] <- predict(predictors, sv[[i]])
    cur.sv.bin[[i]] <- cur.sv[[i]] > svthres[[i]]

    future_variable.sv[[i]] <- predict(future_variable, sv[[i]])
    future_variable.sv.bin[[i]] <- future_variable.sv[[i]] > svthres[[i]]

    future_variable2.sv[[i]] <- predict(future_variable2, sv[[i]])
    future_variable2.sv.bin[[i]] <- future_variable2.sv[[i]] > svthres[[i]]
    
    future_variable3.sv[[i]] <- predict(future_variable3, sv[[i]])
    future_variable3.sv.bin[[i]] <- future_variable3.sv[[i]] > svthres[[i]]

  } else {
    cur.sv[[i]] <- NULL
    cur.sv.bin[[i]] <- NULL
    future_variable.sv[[i]] <- NULL
    future_variable.sv.bin[[i]] <- NULL
    future_variable2.sv[[i]] <- NULL
    future_variable2.sv.bin[[i]] <- NULL
    future_variable3.sv[[i]] <- NULL
    future_variable3.sv.bin[[i]] <- NULL
  }
}


##########
cur.sv <- Filter(Negate(is.null), cur.sv)
cur.sv.bin <- Filter(Negate(is.null), cur.sv.bin)

cur.sv.ens <- Reduce('+', cur.sv.bin)
tval <- unique(cur.sv.ens)
tval <- tval[tval != 0]
tval <- median(tval)

if(!is.null(cur.sv.ens)){
  cur.sv.ens.bin <- cur.sv.ens >= tval} else {
    cur.sv.ens.bin <- NULL
  }

##########
future_variable.sv <- Filter(Negate(is.null), future_variable.sv)
future_variable.sv.bin <- Filter(Negate(is.null), future_variable.sv.bin)
future_variable2.sv <- Filter(Negate(is.null), future_variable2.sv)
future_variable2.sv.bin <- Filter(Negate(is.null), future_variable2.sv.bin)
future_variable3.sv <- Filter(Negate(is.null), future_variable3.sv)
future_variable3.sv.bin <- Filter(Negate(is.null), future_variable3.sv.bin)

future_variable.sv.ens <- Reduce('+', future_variable.sv.bin)
tval <- unique(future_variable.sv.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable.sv.ens.bin <- future_variable.sv.ens >= tval

if(!is.null(future_variable.sv.ens)){
  future_variable.sv.ens.bin <- future_variable.sv.ens >= tval} else {
    future_variable.sv.ens.bin <- NULL
  }

future_variable2.sv.ens <- Reduce('+', future_variable2.sv.bin)
tval <- unique(future_variable2.sv.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable2.sv.ens.bin <- future_variable2.sv.ens >= tval

if(!is.null(future_variable2.sv.ens)){
  future_variable2.sv.ens.bin <- future_variable2.sv.ens >= tval} else {
    future_variable2.sv.ens.bin <- NULL
  }

future_variable3.sv.ens <- Reduce('+', future_variable3.sv.bin)
tval <- unique(future_variable3.sv.ens)
tval <- tval[tval != 0]
tval <- median(tval)
future_variable3.sv.ens.bin <- future_variable3.sv.ens >= tval

if(!is.null(future_variable3.sv.ens)){
  future_variable3.sv.ens.bin <- future_variable3.sv.ens >= tval} else {
    future_variable3.sv.ens.bin <- NULL
  }

##########

if(length(cur.sv) != 0) {

for(z in 1:length(cur.sv)){
  adeq = cur.sv[[z]]
  minimo <- min(adeq[], na.rm=T)
  maximo <- max(adeq[], na.rm=T)
  adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
  cur.sv[[z]] <- calc(adeq, adeq_norm)
}


for(z in 1:length(future_variable.sv)){
  adeq = future_variable.sv[[z]]
  minimo <- min(adeq[], na.rm=T)
  maximo <- max(adeq[], na.rm=T)
  adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
  future_variable.sv[[z]] <- calc(adeq, adeq_norm)
}

for(z in 1:length(future_variable2.sv)){
    adeq = future_variable2.sv[[z]]
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable2.sv[[z]] <- calc(adeq, adeq_norm)
}
  
  for(z in 1:length(future_variable3.sv)){
    adeq = future_variable3.sv[[z]]
    minimo <- min(adeq[], na.rm=T)
    maximo <- max(adeq[], na.rm=T)
    adeq_norm <- function(x) {(x-minimo)/(maximo-minimo)}
    future_variable3.sv[[z]] <- calc(adeq, adeq_norm)
  }

x <- stack(cur.sv)
cur.sv.cont <- calc(x, fun = mean)

x <- stack(future_variable.sv)
future_variable.sv.cont <- calc(x, fun = mean)

x <- stack(future_variable2.sv)
future_variable2.sv.cont <- calc(x, fun = mean)

x <- stack(future_variable3.sv)
future_variable3.sv.cont <- calc(x, fun = mean)

} else {
  cur.sv.cont <- NULL
  future_variable.sv.cont <- NULL
  future_variable2.sv.cont <- NULL
  future_variable3.sv.cont <- NULL
}



###################### ensemble current ALL ALGO
# binarios
cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Projecting ensemble models of', sp.n, '...', '\n')

##########aqui ver de add um ifelse caso algum algoritmo tenha dado abaixo do TSS, que vai dar erro - ignorar se dar zero (NULL)

algo.cur <- list(cur.bc.ens.bin, cur.gm.ens.bin, cur.rf.ens.bin, cur.mx.ens.bin, cur.sv.ens.bin)
algo.cur <- algo.cur[lapply(algo.cur, length) > 0] ##ignores NULL values
algo.cur2 <- algo.cur[!is.na(algo.cur)]
ens.cur <- Reduce('+', algo.cur2)
tval <- unique(ens.cur)
tval <- tval[tval != 0]
tval <- median(tval)
ens.cur.bin <- ens.cur >= tval  ##ensemble presente binario


# Ensemble ----------------------------------------------------------------

# For continuous model, by TSS-weighted average

w <- c(bcTSSval, gmTSSval, rfTSSval, mxTSSval, svTSSval) ##ja ignora o que for null

if (length(cur.bc) != 0) {
st1 <- stack(cur.bc)
for(i in 1:length(cur.bc)){
  min_max <- range(cur.bc[[i]][], na.rm=T)
  #write.table(min_max, paste("cur.bc", i, sp.n, ".txt",  sep="_"))
}} else {
  st1 <- NULL
  }

if (length(cur.gm) != 0){
st2 <- stack(cur.gm)
for(i in 1:length(cur.gm)){
  min_max <- range(cur.gm[[i]][], na.rm=T)
  #write.table(min_max, paste("cur.gm", i, sp.n, ".txt",  sep="_"))
}
#rm(cur.gm)
} else {
  st2 <- NULL
}

if (length(cur.rf) != 0){
st3 <- stack(cur.rf)
for(i in 1:length(cur.rf)){
  min_max <- range(cur.rf[[i]][], na.rm=T)
  #write.table(min_max, paste("cur.rf", i, sp.n, ".txt",  sep="_"))
}
} else {
  st3 <- NULL
}

if (length(cur.mx) != 0){
st4 <- stack(cur.mx)
for(i in 1:length(cur.mx)){
  min_max <- range(cur.mx[[i]][], na.rm=T)
  #write.table(min_max, paste("cur.mx", i, sp.n, ".txt",  sep="_"))
}
#rm(cur.mx)
} else {
  st4 <- NULL
}

if (length(cur.sv) != 0){
st5 <- stack(cur.sv)
for(i in 1:length(cur.sv)){
  min_max <- range(cur.sv[[i]][], na.rm=T)
  #write.table(min_max, paste("cur.sv", i, sp.n, ".txt",  sep="_"))
}
#rm(cur.sv)
} else {
  st5 <- NULL
}

stacks <- c(st1, st2, st3, st4, st5) ## ignora o que for null
st <- stack(stacks)

w.sem.os.nulos <- w[w >= tss.lim]
ens.cur <- weighted.mean(st, w.sem.os.nulos) ##ensamble present continue --resultado final
ens.cur.sd.w <- sum(w.sem.os.nulos * (st - ens.cur)^2) ## variancia
ens.cur.sd.w <- sqrt(ens.cur.sd.w) ##desvio padrão (incerteza) -- pra caso revisor peça


# Uncertainty for each algo (each partition) -------------------------------------------------------------

wbc <- c(bcTSSval)
wbc.sem.os.nulos <- wbc[wbc >= tss.lim]

if(length(wbc.sem.os.nulos) == 0) {
 print('bioclim is null')
  cur.bc.mean.w <- NULL
  cur.bc.sd.w <- NULL
  }else {
cur.bc.mean.w <- weighted.mean(st1, wbc.sem.os.nulos)
cur.bc.sd.w <- sum(wbc.sem.os.nulos * (st1 - cur.bc.mean.w)^2)
cur.bc.sd.w <- sqrt(cur.bc.sd.w) }


wgm <- c(gmTSSval)
wgm.sem.os.nulos <- wgm[wgm >= tss.lim]
if(length(wgm.sem.os.nulos)==0){
  print('glm is null')
  cur.gm.mean.w <- NULL
  cur.gm.sd.w <- NULL
  } else {
cur.gm.mean.w <- weighted.mean(st2, wgm.sem.os.nulos)
cur.gm.sd.w <- sum(wgm.sem.os.nulos * (st2 - cur.gm.mean.w)^2)
cur.gm.sd.w <- sqrt(cur.gm.sd.w)
  }

wrf <- c(rfTSSval)
wrf.sem.os.nulos <- wrf[wrf >= tss.lim]
if (length(wrf.sem.os.nulos)==0){
  print('rf is null')
  cur.rf.mean.w <- NULL
  cur.rf.sd.w <- NULL
  } else {
cur.rf.mean.w <- weighted.mean(st3, wrf.sem.os.nulos)
cur.rf.sd.w <- sum(wrf.sem.os.nulos * (st3 - cur.rf.mean.w)^2)
cur.rf.sd.w <- sqrt(cur.rf.sd.w)
  }

wmx <- c(mxTSSval)
wmx.sem.os.nulos <- wmx[wmx >= tss.lim]
if (length(wmx.sem.os.nulos)==0) {
  print('maxent is null')
  cur.mx.mean.w <- NULL
  cur.mx.sd.w <- NULL
} else {
cur.mx.mean.w <- weighted.mean(st4, wmx.sem.os.nulos)
cur.mx.sd.w <- sum(wmx.sem.os.nulos * (st4 - cur.mx.mean.w)^2)
cur.mx.sd.w <- sqrt(cur.mx.sd.w)
}

wsv <- c(svTSSval)
wsv.sem.os.nulos <- wsv[wsv >= tss.lim]
if (length(wsv.sem.os.nulos)==0){
  print('svmk is null')
  cur.sv.mean.w <- NULL
  cur.sv.sd.w <-  NULL
} else {
cur.sv.mean.w <- weighted.mean(st5, wsv.sem.os.nulos)
cur.sv.sd.w <- sum(wsv.sem.os.nulos * (st5 - cur.sv.mean.w)^2)
cur.sv.sd.w <- sqrt(cur.sv.sd.w)
}

# Ensemble (future) GCM 1 -------------------------------------------------------------

# For binary model

# binarios
algo.future_variable <- c(future_variable.bc.ens.bin, future_variable.gm.ens.bin, future_variable.rf.ens.bin, future_variable.mx.ens.bin, future_variable.sv.ens.bin)
algo.future_variable <- algo.future_variable[lapply(algo.future_variable, length) > 0] ##ignores NULL rasters but not NA

ens.future_variable <- Reduce('+', algo.future_variable)
tval <- unique(ens.future_variable)
tval <- tval[tval != 0]
tval <- median(tval)
ens.future_variable.bin <- ens.future_variable >= tval



if (length(future_variable.bc) != 0){
st1 <- stack(future_variable.bc)
for(i in 1:length(future_variable.bc)){
  min_max <- range(future_variable.bc[[i]][], na.rm=T)
  #write.table(min_max, paste("future_variable.bc", i, sp.n, ".txt",  sep="_"))
}} else {
  st1 <- NULL
}
#rm(future_variable.bc)

# continuous com ifelse

if (length(future_variable.gm) != 0){
st2 <- stack(future_variable.gm)
for(i in 1:length(future_variable.gm)){
  min_max <- range(future_variable.gm[[i]][], na.rm=T)
  #write.table(min_max, paste("future_variable.gm", i, sp.n, ".txt",  sep="_"))
}} else{
  st2 <- NULL
}


if (length(future_variable.rf) != 0){
st3 <- stack(future_variable.rf)
for(i in 1:length(future_variable.rf)){
  min_max <- range(future_variable.rf[[i]][], na.rm=T)
  #write.table(min_max, paste("future_variable.rf", i, sp.n, ".txt",  sep="_"))
}} else {
  st3 <- NULL
}


if (length(future_variable.mx) != 0){
st4 <- stack(future_variable.mx)
for(i in 1:length(future_variable.mx)){
  min_max <- range(future_variable.mx[[i]][], na.rm=T)
  #write.table(min_max, paste("future_variable.mx", i, sp.n, ".txt",  sep="_"))
}} else {
st4 <- NULL
  }



if (length(future_variable.sv) != 0){
st5 <- stack(future_variable.sv)
for(i in 1:length(future_variable.sv)){
  min_max <- range(future_variable.sv[[i]][], na.rm=T)
  #write.table(min_max, paste("future_variable.sv", i, sp.n, ".txt",  sep="_"))
}} else {
  st5 <- NULL
}


#rm(future_variable.sv)

stack_fut <- c(st1, st2, st3, st4, st5)
st_fut <- stack(stack_fut)
w.sem.os.nulos <- w[w >= tss.lim]
ens2.future_variable <- weighted.mean(st_fut, w.sem.os.nulos)
ens.fut.cont <- ens2.future_variable
ens2.fut.sd.w <- sum(w.sem.os.nulos * (st_fut - ens.fut.cont)^2)
ens2.fut.sd.w <- sqrt(ens2.fut.sd.w)


# Uncertainty future ------------------------------------------------------

wbc <- c(bcTSSval)
wbc.sem.os.nulos <- wbc[wbc >= tss.lim]
if(length(wbc.sem.os.nulos) == 0){
  print('bioclim is null')
  future_variable.bc.mean.w <- NULL
  future_variable.bc.sd.w <-  NULL
  } else {
future_variable.bc.mean.w <- weighted.mean(st1, wbc.sem.os.nulos)
future_variable.bc.sd.w <- sum(wbc.sem.os.nulos * (st1 - future_variable.bc.mean.w)^2)
future_variable.bc.sd.w <- sqrt(future_variable.bc.sd.w) }

wgm <- c(gmTSSval)
wgm.sem.os.nulos <- wgm[wgm >= tss.lim]
if(length(wgm.sem.os.nulos) == 0){
  print('glm is null')
  future_variable.gm.mean.w <- NULL
  future_variable.gm.sd.w <- NULL
  } else {
future_variable.gm.mean.w <- weighted.mean(st2, wgm.sem.os.nulos)
future_variable.gm.sd.w <- sum(wgm.sem.os.nulos * (st2 - future_variable.gm.mean.w)^2)
future_variable.gm.sd.w <- sqrt(future_variable.gm.sd.w)}

wrf <- c(rfTSSval)
wrf.sem.os.nulos <- wrf[wrf >= tss.lim]
if(length(wrf.sem.os.nulos) == 0){
  print('rf is null')
  future_variable.rf.mean.w <- NULL
  future_variable.rf.sd.w <- NULL
  } else {
future_variable.rf.mean.w <- weighted.mean(st3, wrf.sem.os.nulos)
future_variable.rf.sd.w <- sum(wrf.sem.os.nulos * (st3 - future_variable.rf.mean.w)^2)
future_variable.rf.sd.w <- sqrt(future_variable.rf.sd.w)}

wmx <- c(mxTSSval)
wmx.sem.os.nulos <- wmx[wmx >= tss.lim]
if(length(wmx.sem.os.nulos) == 0){
  print('mx is null')
  future_variable.mx.mean.w <- NULL
  future_variable.mx.sd.w <- NULL
    } else {
future_variable.mx.mean.w <- weighted.mean(st4, wmx.sem.os.nulos)
future_variable.mx.sd.w <- sum(wmx.sem.os.nulos * (st4 - future_variable.mx.mean.w)^2)
future_variable.mx.sd.w <- sqrt(future_variable.mx.sd.w)}

wsv <- c(svTSSval)
wsv.sem.os.nulos <- wsv[wsv >= tss.lim]
if(length(wsv.sem.os.nulos) == 0){
  print('svm is null')
  future_variable.sv.mean.w <- NULL
  future_variable.sv.sd.w <- NULL
  } else {
future_variable.sv.mean.w <- weighted.mean(st5, wsv.sem.os.nulos)
future_variable.sv.sd.w <- sum(wsv.sem.os.nulos * (st5 - future_variable.sv.mean.w)^2)
future_variable.sv.sd.w <- sqrt(future_variable.sv.sd.w)}

####

ens.fut <- ens.future_variable.bin
tval <- unique(ens.fut)
tval <- tval[tval != 0]
tval <- median(tval)
ens.fut.bin <- ens.fut >= tval

# Ensemble (future) GCM 2 -------------------------------------------------------------

# For binary model

# binarios
algo.future_variable2 <- c(future_variable2.bc.ens.bin, future_variable2.gm.ens.bin, future_variable2.rf.ens.bin, future_variable2.mx.ens.bin, future_variable2.sv.ens.bin)
algo.future_variable2 <- algo.future_variable2[lapply(algo.future_variable2, length) > 0] ##ignores NULL rasters but not NA
#algo.future_variable2 <- list()

## Ignores rasters that are all NA
#for (i in 1:length(algo.future_variable)){
#if (!is.na(minValue(algo.future_variable[[i]]))){
# print('it is not NA')
# algo.future_variable2[[i]] <- algo.future_variable[[i]]
#}
#}

ens.future_variable2 <- Reduce('+', algo.future_variable2)
tval <- unique(ens.future_variable2)
tval <- tval[tval != 0]
tval <- median(tval)
ens.future_variable2.bin <- ens.future_variable2 >= tval



if (length(future_variable2.bc) != 0){
  st1 <- stack(future_variable2.bc)
  for(i in 1:length(future_variable2.bc)){
    min_max <- range(future_variable2.bc[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.bc", i, sp.n, ".txt",  sep="_"))
  }} else {
    st1 <- NULL
  }
#rm(future_variable.bc)

# continuous com ifelse

if (length(future_variable2.gm) != 0){
  st2 <- stack(future_variable2.gm)
  for(i in 1:length(future_variable2.gm)){
    min_max <- range(future_variable2.gm[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.gm", i, sp.n, ".txt",  sep="_"))
  }} else{
    st2 <- NULL
  }


if (length(future_variable2.rf) != 0){
  st3 <- stack(future_variable2.rf)
  for(i in 1:length(future_variable2.rf)){
    min_max <- range(future_variable2.rf[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.rf", i, sp.n, ".txt",  sep="_"))
  }} else {
    st3 <- NULL
  }


if (length(future_variable2.mx) != 0){
  st4 <- stack(future_variable2.mx)
  for(i in 1:length(future_variable2.mx)){
    min_max <- range(future_variable2.mx[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.mx", i, sp.n, ".txt",  sep="_"))
  }} else {
    st4 <- NULL
  }



if (length(future_variable2.sv) != 0){
  st5 <- stack(future_variable2.sv)
  for(i in 1:length(future_variable2.sv)){
    min_max <- range(future_variable2.sv[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.sv", i, sp.n, ".txt",  sep="_"))
  }} else {
    st5 <- NULL
  }


#rm(future_variable.sv)

stack_fut2 <- c(st1, st2, st3, st4, st5)
st_fut2 <- stack(stack_fut2)
w.sem.os.nulos <- w[w >= tss.lim]
ens2.future_variable2 <- weighted.mean(st_fut2, w.sem.os.nulos)
ens.fut.cont2 <- ens2.future_variable2
ens2.fut.sd.w2 <- sum(w.sem.os.nulos * (st_fut2 - ens.fut.cont2)^2)
ens2.fut.sd.w2 <- sqrt(ens2.fut.sd.w2)


# Uncertainty future ------------------------------------------------------

wbc <- c(bcTSSval)
wbc.sem.os.nulos <- wbc[wbc >= tss.lim]
if(length(wbc.sem.os.nulos) == 0){
  print('bioclim is null')
  future_variable2.bc.mean.w <- NULL
  future_variable2.bc.sd.w <- NULL
} else {
  future_variable2.bc.mean.w <- weighted.mean(st1, wbc.sem.os.nulos)
  future_variable2.bc.sd.w <- sum(wbc.sem.os.nulos * (st1 - future_variable2.bc.mean.w)^2)
  future_variable2.bc.sd.w <- sqrt(future_variable2.bc.sd.w) }

wgm <- c(gmTSSval)
wgm.sem.os.nulos <- wgm[wgm >= tss.lim]
if(length(wgm.sem.os.nulos) == 0){
  print('glm is null')
  future_variable2.gm.mean.w <- NULL
    future_variable2.gm.sd.w <- NULL
} else {
  future_variable2.gm.mean.w <- weighted.mean(st2, wgm.sem.os.nulos)
  future_variable2.gm.sd.w <- sum(wgm.sem.os.nulos * (st2 - future_variable2.gm.mean.w)^2)
  future_variable2.gm.sd.w <- sqrt(future_variable2.gm.sd.w)}

wrf <- c(rfTSSval)
wrf.sem.os.nulos <- wrf[wrf >= tss.lim]
if(length(wrf.sem.os.nulos) == 0){
  print('rf is null')
  future_variable2.rf.mean.w <- NULL
  future_variable2.rf.sd.w <- NULL
} else {
  future_variable2.rf.mean.w <- weighted.mean(st3, wrf.sem.os.nulos)
  future_variable2.rf.sd.w <- sum(wrf.sem.os.nulos * (st3 - future_variable2.rf.mean.w)^2)
  future_variable2.rf.sd.w <- sqrt(future_variable2.rf.sd.w)}

wmx <- c(mxTSSval)
wmx.sem.os.nulos <- wmx[wmx >= tss.lim]
if(length(wmx.sem.os.nulos) == 0){
  print('mx is null')
  future_variable2.mx.mean.w <- NULL
  future_variable2.mx.sd.w <- NULL
} else {
  future_variable2.mx.mean.w <- weighted.mean(st4, wmx.sem.os.nulos)
  future_variable2.mx.sd.w <- sum(wmx.sem.os.nulos * (st4 - future_variable2.mx.mean.w)^2)
  future_variable2.mx.sd.w <- sqrt(future_variable2.mx.sd.w)}

wsv <- c(svTSSval)
wsv.sem.os.nulos <- wsv[wsv >= tss.lim]
if(length(wsv.sem.os.nulos) == 0){
  print('svm is null')
  future_variable2.sv.mean.w <- NULL
  future_variable2.sv.sd.w <- NULL
} else {
  future_variable2.sv.mean.w <- weighted.mean(st5, wsv.sem.os.nulos)
  future_variable2.sv.sd.w <- sum(wsv.sem.os.nulos * (st5 - future_variable2.sv.mean.w)^2)
  future_variable2.sv.sd.w <- sqrt(future_variable2.sv.sd.w)}

####

ens.fut2 <- ens.future_variable2.bin
tval <- unique(ens.fut2)
tval <- tval[tval != 0]
tval <- median(tval)
ens.fut.bin2 <- ens.fut2 >= tval

# Ensemble (future) GCM3 -------------------------------------------------------------

# For binary model

# binarios
algo.future_variable3 <- c(future_variable3.bc.ens.bin, future_variable3.gm.ens.bin, future_variable3.rf.ens.bin, future_variable3.mx.ens.bin, future_variable3.sv.ens.bin)
algo.future_variable3 <- algo.future_variable3[lapply(algo.future_variable3, length) > 0] ##ignores NULL rasters but not NA
#algo.future_variable2 <- list()

## Ignores rasters that are all NA
#for (i in 1:length(algo.future_variable)){
#if (!is.na(minValue(algo.future_variable[[i]]))){
# print('it is not NA')
# algo.future_variable2[[i]] <- algo.future_variable[[i]]
#}
#}

ens.future_variable3 <- Reduce('+', algo.future_variable3)
tval <- unique(ens.future_variable3)
tval <- tval[tval != 0]
tval <- median(tval)
ens.future_variable3.bin <- ens.future_variable3 >= tval



if (length(future_variable3.bc) != 0){
  st1 <- stack(future_variable3.bc)
  for(i in 1:length(future_variable3.bc)){
    min_max <- range(future_variable3.bc[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.bc", i, sp.n, ".txt",  sep="_"))
  }} else {
    st1 <- NULL
  }
#rm(future_variable.bc)

# continuous com ifelse

if (length(future_variable3.gm) != 0){
  st2 <- stack(future_variable3.gm)
  for(i in 1:length(future_variable3.gm)){
    min_max <- range(future_variable3.gm[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.gm", i, sp.n, ".txt",  sep="_"))
  }} else{
    st2 <- NULL
  }


if (length(future_variable3.rf) != 0){
  st3 <- stack(future_variable3.rf)
  for(i in 1:length(future_variable3.rf)){
    min_max <- range(future_variable3.rf[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.rf", i, sp.n, ".txt",  sep="_"))
  }} else {
    st3 <- NULL
  }


if (length(future_variable3.mx) != 0){
  st4 <- stack(future_variable3.mx)
  for(i in 1:length(future_variable3.mx)){
    min_max <- range(future_variable3.mx[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.mx", i, sp.n, ".txt",  sep="_"))
  }} else {
    st4 <- NULL
  }



if (length(future_variable3.sv) != 0){
  st5 <- stack(future_variable3.sv)
  for(i in 1:length(future_variable3.sv)){
    min_max <- range(future_variable3.sv[[i]][], na.rm=T)
    #write.table(min_max, paste("future_variable.sv", i, sp.n, ".txt",  sep="_"))
  }} else {
    st5 <- NULL
  }


#rm(future_variable.sv)

stack_fut3 <- c(st1, st2, st3, st4, st5)
st_fut3 <- stack(stack_fut3)
w.sem.os.nulos <- w[w >= tss.lim]
ens2.future_variable3 <- weighted.mean(st_fut3, w.sem.os.nulos)
ens.fut.cont3 <- ens2.future_variable3
ens2.fut.sd.w3 <- sum(w.sem.os.nulos * (st_fut3 - ens.fut.cont3)^2)
ens2.fut.sd.w3 <- sqrt(ens2.fut.sd.w3)


# Uncertainty future ------------------------------------------------------

wbc <- c(bcTSSval)
wbc.sem.os.nulos <- wbc[wbc >= tss.lim]
if(length(wbc.sem.os.nulos) == 0){
  print('bioclim is null')
  future_variable3.bc.mean.w <- NULL
  future_variable3.bc.sd.w <- NULL
} else {
  future_variable3.bc.mean.w <- weighted.mean(st1, wbc.sem.os.nulos)
  future_variable3.bc.sd.w <- sum(wbc.sem.os.nulos * (st1 - future_variable3.bc.mean.w)^2)
  future_variable3.bc.sd.w <- sqrt(future_variable3.bc.sd.w) }

wgm <- c(gmTSSval)
wgm.sem.os.nulos <- wgm[wgm >= tss.lim]
if(length(wgm.sem.os.nulos) == 0){
  print('glm is null')
  future_variable3.gm.mean.w <- NULL
  future_variable3.gm.sd.w <- NULL
} else {
  future_variable3.gm.mean.w <- weighted.mean(st2, wgm.sem.os.nulos)
  future_variable3.gm.sd.w <- sum(wgm.sem.os.nulos * (st2 - future_variable3.gm.mean.w)^2)
  future_variable3.gm.sd.w <- sqrt(future_variable3.gm.sd.w)}

wrf <- c(rfTSSval)
wrf.sem.os.nulos <- wrf[wrf >= tss.lim]
if(length(wrf.sem.os.nulos) == 0){
  print('rf is null')
  future_variable3.rf.mean.w <- NULL
  future_variable3.rf.sd.w <- NULL
} else {
  future_variable3.rf.mean.w <- weighted.mean(st3, wrf.sem.os.nulos)
  future_variable3.rf.sd.w <- sum(wrf.sem.os.nulos * (st3 - future_variable3.rf.mean.w)^2)
  future_variable3.rf.sd.w <- sqrt(future_variable3.rf.sd.w)}

wmx <- c(mxTSSval)
wmx.sem.os.nulos <- wmx[wmx >= tss.lim]
if(length(wmx.sem.os.nulos) == 0){
  print('mx is null')
  future_variable3.mx.mean.w <- NULL
  future_variable3.mx.sd.w <- NULL
} else {
  future_variable3.mx.mean.w <- weighted.mean(st4, wmx.sem.os.nulos)
  future_variable3.mx.sd.w <- sum(wmx.sem.os.nulos * (st4 - future_variable3.mx.mean.w)^2)
  future_variable3.mx.sd.w <- sqrt(future_variable3.mx.sd.w)}

wsv <- c(svTSSval)
wsv.sem.os.nulos <- wsv[wsv >= tss.lim]
if(length(wsv.sem.os.nulos) == 0){
  print('svm is null')
  future_variable3.sv.mean.w <- NULL
  future_variable3.sv.sd.w <- NULL
} else {
  future_variable3.sv.mean.w <- weighted.mean(st5, wsv.sem.os.nulos)
  future_variable3.sv.sd.w <- sum(wsv.sem.os.nulos * (st5 - future_variable3.sv.mean.w)^2)
  future_variable3.sv.sd.w <- sqrt(future_variable3.sv.sd.w)}

####

ens.fut3 <- ens.future_variable3.bin
tval <- unique(ens.fut3)
tval <- tval[tval != 0]
tval <- median(tval)
ens.fut.bin3 <- ens.fut3 >= tval



# Ensemble GCMs

#Binary
ens_gcm_bin <- c(ens.future_variable.bin + ens.future_variable2.bin + ens.future_variable3.bin)
ens_gcm_bin <- ens_gcm_bin[lapply(ens_gcm_bin, length) > 0]

ens_gcm_all_bin <- Reduce('+', ens_gcm_bin)
tval <- unique(ens_gcm_all_bin)
tval <- tval[tval != 0]
tval <- median(tval)
ens_gcm_all_bin <- ens_gcm_all_bin >= tval

#Continuous
ens_gcm_cont <- (ens.fut.cont + ens.fut.cont2 + ens.fut.cont3)/3


###SAVING

cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Saving ensemble maps of', sp.n, '...', '\n')
if(!is.null(ens.cur)){
writeRaster(ens.cur, file = paste(target_dir, '/CUR.cont_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens2.future_variable, file = paste(target_dir, '/Fut_bcc.cont_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens2.future_variable2, file = paste(target_dir, '/Fut_ipsl.cont_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens2.future_variable3, file = paste(target_dir, '/Fut_miroc.cont_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens_gcm_cont, file = paste(target_dir, '/Fut_all.cont_', sp.n, '.asc', sep=""),overwrite=TRUE)

}
if(!is.null(ens.cur.bin)){
writeRaster(ens.cur.bin, file = paste(target_dir, '/CUR.bin_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens.future_variable.bin, file = paste(target_dir, '/Fut_bcc.bin_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens.future_variable2.bin, file = paste(target_dir, '/Fut_ipsl.bin_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens.future_variable3.bin, file = paste(target_dir, '/Fut_miroc.bin_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens_gcm_all_bin, file = paste(target_dir, '/Fut_all.bin_', sp.n, '.asc', sep=""),overwrite=TRUE)

}

#weighted.mean and weighted.sd
if(!is.null(ens.cur.sd.w)){
writeRaster(ens.cur.sd.w, file = paste(target_dir, 'ens.cur.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens2.fut.sd.w, file = paste(target_dir, 'ens.Fut_bcc.', BCC, '.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens2.fut.sd.w2, file = paste(target_dir, 'ens.Fut_ipsl.', IPSL, '.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
writeRaster(ens2.fut.sd.w3, file = paste(target_dir, 'ens.Fut_miroc.', MIROC, '.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)

}
#writeRaster(cur.bc.mean.w, file = paste(target_dir, '/cur.bc.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
#writeRaster(cur.bc.sd.w, file = paste(target_dir, '/cur.bc.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
#writeRaster(cur.gm.mean.w, file = paste(target_dir, '/cur.gm.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
if(!is.null(cur.gm.sd.w)){
writeRaster(cur.gm.sd.w, file = paste(target_dir, 'cur.gm.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}
  #writeRaster(cur.rf.mean.w, file = paste(target_dir, '/cur.rf.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
if(!is.null(cur.rf.sd.w)){
writeRaster(cur.rf.sd.w, file = paste(target_dir, 'cur.rf.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}
  #writeRaster(cur.mx.mean.w, file = paste(target_dir, '/cur.mx.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
if(!is.null(cur.mx.sd.w)){
writeRaster(cur.mx.sd.w, file = paste(target_dir, 'cur.mx.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}
  #writeRaster(cur.sv.mean.w, file = paste(target_dir, '/cur.sv.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
if(!is.null(cur.sv.sd.w)){
writeRaster(cur.sv.sd.w, file = paste(target_dir, 'cur.sv.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

#writeRaster(future_variable.bc.mean.w, file = paste(target_dir, '/mpi8570.bc.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
#writeRaster(future_variable.bc.sd.w, file = paste(target_dir, '/mpi8570.bc.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
#writeRaster(future_variable.gm.mean.w, file = paste(target_dir, '/mpi8570.gm.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
if(!is.null(future_variable.gm.sd.w)){
writeRaster(future_variable.gm.sd.w, file = paste(target_dir, 'Fut_', BCC, '.gm.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

if(!is.null(future_variable2.gm.sd.w)){
  writeRaster(future_variable2.gm.sd.w, file = paste(target_dir, 'Fut_', IPSL, '.gm.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}
  #writeRaster(future_variable.rf.mean.w, file = paste(target_dir, '/mpi8570.rf.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)

if(!is.null(future_variable3.gm.sd.w)){
  writeRaster(future_variable3.gm.sd.w, file = paste(target_dir, 'Fut_', MIROC, '.gm.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}


if(!is.null(future_variable.rf.sd.w)){
writeRaster(future_variable.rf.sd.w, file = paste(target_dir, '/Fut_', BCC, '.rf.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

if(!is.null(future_variable2.rf.sd.w)){
  writeRaster(future_variable2.rf.sd.w, file = paste(target_dir, '/Fut_', IPSL, '.rf.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

if(!is.null(future_variable3.rf.sd.w)){
  writeRaster(future_variable3.rf.sd.w, file = paste(target_dir, '/Fut_', MIROC, '.rf.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

  #writeRaster(future_variable.mx.mean.w, file = paste(target_dir, '/mpi8570.mx.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
if(!is.null(future_variable.mx.sd.w)){
writeRaster(future_variable.mx.sd.w, file = paste(target_dir, '/Fut_', BCC, '.mx.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

if(!is.null(future_variable2.mx.sd.w)){
  writeRaster(future_variable2.mx.sd.w, file = paste(target_dir, '/Fut_', IPSL, '.mx.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

if(!is.null(future_variable3.mx.sd.w)){
  writeRaster(future_variable3.mx.sd.w, file = paste(target_dir, '/Fut_', MIROC, '.mx.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

  #writeRaster(future_variable.sv.mean.w, file = paste(target_dir, '/mpi8570.sv.mean.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
if(!is.null(future_variable.sv.sd.w)){
writeRaster(future_variable.sv.sd.w, file = paste(target_dir, '/Fut_', BCC, '.sv.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

if(!is.null(future_variable2.sv.sd.w)){
  writeRaster(future_variable2.sv.sd.w, file = paste(target_dir, '/Fut_', IPSL, '.sv.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

if(!is.null(future_variable3.sv.sd.w)){
  writeRaster(future_variable3.sv.sd.w, file = paste(target_dir, '/Fut_', MIROC, '.sv.sd.w_', sp.n, '.asc', sep=""),overwrite=TRUE)
}

cat( format( Sys.time(), "%a %b %d %X %Y"), '-', 'Finished train and test datasets for', sp.n, 'with ', lim, 'lines...', '\n')

#save.image("./outputs/my_analysis.rData")

finished_time = Sys.time()
cat( format( finished_time, "%a %b %d %X %Y"), '-', 'FINISHED', '\n')
write(format( finished_time, "%a %b %d %X %Y"), file = paste0(target_dir, "FINISHED.txt", sep=""))

}

#beep(5)
